requirements.txt
boto3==1.28.63
botocore==1.31.63
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
jmespath==1.0.1
mojito2==0.1.6
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
mysql==0.0.3
mysql-connector-python==8.1.0
mysqlclient==2.1.1
numpy==1.26.0
pandas==2.1.1
protobuf==4.21.12
pycryptodome==3.19.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
SQLAlchemy==2.0.22
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
websockets==11.0.3

